% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{add_profile_pics_to_tree_ggraph}
\alias{add_profile_pics_to_tree_ggraph}
\title{Create ggraph object with profile pictures added}
\usage{
add_profile_pics_to_tree_ggraph(g, df_profile_pic)
}
\arguments{
\item{g}{tbl_graph object}

\item{df_profile_pic}{tibble of profile pictures, generated by `get_profile_pic_df()`}
}
\value{
ggraph object of tweet tree with profile pictures
}
\description{
Create ggraph object with profile pictures added
}
\examples{
\dontrun{
g1 <- add_profile_pics_to_tree_ggraph(g, get_profile_pic_df(bind_rows(df_tls, df_main_status)))
g1 + ggraph::geom_edge_diagonal(aes(color= type)) + geom_node_point(shape = type))
}
}
