% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{add_tree_level}
\alias{add_tree_level}
\title{Add one level in the tree of tweets}
\usage{
add_tree_level(df0, df1, n)
}
\arguments{
\item{df0}{data frame.}

\item{df1}{data frame.}

\item{n}{maximum number of tweets to scrape.}
}
\value{
data frame like `rtweet::search_tweets2()`, but also all direct answers
  to the tweet (one level lower; if available).
}
\description{
Add one level in the tree of tweets
}
\keyword{internal}
