% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweettree.R
\name{autoplot.character}
\alias{autoplot.character}
\title{Plot rtweettree}
\usage{
\method{autoplot}{character}(x, ...)
}
\arguments{
\item{x}{rtweet status_id or rtweettree_data object}

\item{...}{for the moment not used}
}
\value{
rtweettree
}
\description{
Plot rtweettree
}
\examples{
main_status_id <- "1438481824922181635"
\dontrun{
l <- rtweettree_data(main_status_id)
autoplot(l)
}
# when you're not interested to store the scraped data in an R object,
# you can also directly plot it with:
# autoplot(main_status_id)

}
