% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweettree_plot.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot.character}
\alias{autoplot.rtweettree_data}
\alias{autoplot.list}
\alias{autoplot.rtweettree_tbl_graph}
\title{Plot rtweettree}
\usage{
\method{autoplot}{character}(x, add_profile_pics = TRUE, ...)

\method{autoplot}{rtweettree_data}(x, add_profile_pics = TRUE, ...)

\method{autoplot}{list}(x, add_profile_pics = TRUE, ...)

\method{autoplot}{rtweettree_tbl_graph}(x, add_profile_pics = TRUE, ...)
}
\arguments{
\item{x}{rtweet status_id (character string), rtweettree_data or rtweettree_tbl_graph object}

\item{add_profile_pics}{logical whether to add the profile pictures of the
users to the graph; defaults to TRUE; (should be set to FALSE for large graphs)}

\item{...}{arguments passed to methods}
}
\value{
rtweettree graph
}
\description{
Plot a tree graph of the data resulting of the `status_id` of a tweet.
}
\details{
The following functions are imported and then re-exported
from the ggplot2 package to avoid loading them.

rtweettree exported operators and S3 methods
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: The `status_id` charcter string is transformed to a rtweettree_tbl_graph which is then plotted with ggraph.

\item \code{rtweettree_data}: The rtweettree_data object is transformed to a rtweettree_tbl_graph which is then plotted with ggraph.

\item \code{list}: The list is transformed to a rtweettree_tbl_graph which is then plotted with ggraph.

\item \code{rtweettree_tbl_graph}: The rtweettree_tbl_graph is plotted with ggraph.
}}

\examples{
main_status_id <- "1438481824922181635"
\dontrun{
l <- rtweettree_data(main_status_id)
autoplot(l)
}
# when you're not interested to store the scraped data in an R object,
# you can also directly plot it with:
# autoplot(main_status_id)

}
