% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_rtweet_tbl}
\alias{create_empty_rtweet_tbl}
\title{Create an empty rtweet tibble}
\usage{
create_empty_rtweet_tbl()
}
\value{
An empty tibble with columns of the type that, e.g. rtweet::lookup_statuses() produces
}
\description{
Create an empty rtweet tibble
}
\examples{
df <- rtweettree:::create_empty_rtweet_tbl()
df
}
\keyword{internal}
