% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_favs}
\alias{df_favs}
\title{Example package dataset storing the result of scrape_favs2(ids, main_status_id)}
\format{
A data frame with 1 row and 91 variables:
\describe{
  df_favs <- scrape_favs2(ids, main_status_id)
}
}
\usage{
df_favs
}
\description{
Example package dataset storing the result of scrape_favs2(ids, main_status_id)
}
\keyword{datasets}
