% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_profile_pic}
\alias{df_profile_pic}
\title{Example package dataset storing the result of rtweettree:::get_profile_pic_df()}
\format{
A data frame with 1 row and 90 variables:
\describe{
  df_profile_pic <- rtweettree:::get_profile_pic_df(bind_rows(df_tls, df_favs, df_main_status))
}
}
\usage{
df_profile_pic
}
\description{
Example package dataset storing the result of rtweettree:::get_profile_pic_df()
}
\keyword{datasets}
