% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_retweets}
\alias{df_retweets}
\title{Example package dataset storing the result of df_retweets <- tweet_ids %>% map_dfr(~rtweet::get_retweets(.x))}
\format{
A data frame with 1 row and 90 variables:
\describe{
  df_retweets <- tweet_ids %>% map_dfr(~rtweet::get_retweets(.x))
}
}
\usage{
df_retweets
}
\description{
Example package dataset storing the result of df_retweets <- tweet_ids %>% map_dfr(~rtweet::get_retweets(.x))
}
\keyword{datasets}
