% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_tls}
\alias{df_tls}
\title{Example package dataset storing the result of rtweet::get_timelines(tree_ids)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 90 columns.
}
\usage{
df_tls
}
\description{
Example package dataset storing the result of rtweet::get_timelines(tree_ids)
}
\examples{
\dontrun{
  df_tls <- rtweet::get_timelines(tree_ids)
}
}
\keyword{datasets}
