% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{find_connections_rec}
\alias{find_connections_rec}
\title{Recursively return all \code{status_id}s of a tweet and its replies}
\usage{
find_connections_rec(df_thread, df0 = df_main_status)
}
\arguments{
\item{df_thread}{data frame returned by rtweet::search_tweets2(main_status_id)}

\item{df_main_status}{data frame returned by rtweet::lookup_statuses(main_status_id)}
}
\value{
Edges data frame that can be plotted by tidygraph & ggraph
}
\description{
Recursively return all \code{status_id}s of a tweet and its replies
}
\examples{
main_status_id <- "1234620900386975744"
df_main_status <- rtweet::lookup_statuses(main_status_id)
df_thread <- search_thread(main_status_id)
thread_ids <- df_thread$user_id \%>\% unique()
df_tls <- scrape_timelines(thread_ids)
df0 <- df_main_status \%>\%
                       dplyr::filter(status_id == main_status_id) \%>\%
                       dplyr::select(to = status_id, user_id) \%>\%
                       dplyr::mutate(from = "root", type = "root")
tweet_edges <-
find_connections_rec(dplyr::bind_rows(df_thread, df_tls), df0)
}
