% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incl_data}
\alias{incl_data}
\alias{rtweettree_data_example}
\title{Included example dataset}
\format{
An object of class \code{rtweettree_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12 rows and 93 columns.
}
\usage{
rtweettree_data_example
}
\description{
The package includes a dataset resulting of rtweet queries when calling
`rtweettree_data()` (to make them available for documentation purposes and
offline use).
}
\examples{
\dontrun{
main_status_id <- "1438481824922181635"
rtweettree_data(main_status_id)
}
rtweettree_data_example
}
\keyword{datasets}
