% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_favs2}
\alias{scrape_favs2}
\title{Scrape all likes of all users occurring in a tree of a twitter status id}
\usage{
scrape_favs2(ids, main_status_id)
}
\arguments{
\item{ids}{Vector of all `user_id`s.}

\item{main_status_id}{status id of the root tweet.}
}
\value{
Dataframe of all timelines of all `tree_ids`.
}
\description{
Scrape all likes of all users occurring in a tree of a twitter status id
}
\examples{
\dontrun{
main_status_id <- "1234620900386975744"
df_main_status <- rtweet::lookup_statuses(main_status_id)
df_tree <- search_tree(main_status_id)
tree_ids <- df_tree$user_id \%>\% unique()
df_tls <- scrape_timelines(tree_ids, main_status_id)
df0 <- df_main_status \%>\%
                       dplyr::filter(status_id == main_status_id) \%>\%
                       dplyr::select(to = status_id, user_id) \%>\%
                       dplyr::mutate(from = "root", type = "root")
tweet_edges <-
find_connections_rec(dplyr::bind_rows(df_tree, df_tls), df0)
ids <- tweet_edges$user_id \%>\% unique()
df_favs <- scrape_favs2(ids, main_status_id)
}
}
