% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{search_thread}
\alias{search_thread}
\title{Scrape thread}
\usage{
search_thread(
  main_status_id,
  df_main_status = rtweet::lookup_statuses(main_status_id),
  save_res = TRUE,
  n = 1e+06
)
}
\arguments{
\item{main_status_id}{string of twitter status_id}

\item{df_main_status}{data frame returned by rtweet::lookup_statuses(main_status_id)}

\item{save_res}{logical if file should be saved}

\item{n}{maximum number of tweets to scrape}
}
\value{
data frame like rtweet::search_tweets2, but all subtweets of the
  thread added (if available)
}
\description{
Scrape thread
}
\examples{
main_status_id <- "1234620900386975744"
df_main_status <- rtweet::lookup_statuses(main_status_id)
df_thread <- search_thread(main_status_id)
}
