% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{search_tree}
\alias{search_tree}
\title{Scrape tree}
\usage{
search_tree(
  main_status_id,
  df_main_status = rtweet::lookup_statuses(main_status_id),
  n = 1e+06
)
}
\arguments{
\item{main_status_id}{string of twitter status_id}

\item{df_main_status}{data frame returned by rtweet::lookup_statuses(main_status_id)}

\item{n}{maximum number of tweets to scrape}
}
\value{
data frame like rtweet::search_tweets2, but all subtweets of the
  tree added (if available)
}
\description{
Scrape tree
}
\examples{
\dontrun{
main_status_id <- "1234620900386975744"
df_main_status <- rtweet::lookup_statuses(main_status_id)
df_tree <- search_tree(main_status_id)
}
}
