% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{add_profile_pics_to_tree_ggraph}
\alias{add_profile_pics_to_tree_ggraph}
\title{Create ggraph object with profile pictures added}
\usage{
add_profile_pics_to_tree_ggraph(g1, df_profile_pic)
}
\arguments{
\item{g1}{tbl_graph object}

\item{df_profile_pic}{tibble of profile pictures, generated by `scrape_profile_pics()` (see examples).}
}
\value{
ggraph object of tweet tree with profile pictures
}
\description{
Create ggraph object with profile pictures added
}
\examples{
\dontrun{
rtweettree_data_scraped <- rtweettree_data("1438481824922181635")
g <- rtweettree_tbl_graph(rtweettree_data_scraped)
g1 <- ggraph(g)
g1 <- add_profile_pics_to_tree_ggraph(g1, rtweettree:::scrape_profile_pics(tidygraph::as_tibble(g)))
g1 + ggraph::geom_edge_diagonal(aes(color= type)) + geom_node_point(aes(shape = type))
}
}
