% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_favs}
\alias{df_favs}
\title{Example package dataset storing the result of scrape_favs2(ids, main_status_id)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 91 columns.
}
\usage{
df_favs
}
\description{
Example package dataset storing the result of scrape_favs2(ids, main_status_id)
}
\examples{
\dontrun{
  df_favs <- scrape_favs2(ids, main_status_id)
}
}
\keyword{datasets}
