% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_main_status}
\alias{df_main_status}
\title{Example package dataset storing the result of rtweet::lookup_statuses(main_status_id)}
\format{
A data frame with 1 row and 90 variables:
\describe{
  df_main_status <- rtweet::lookup_statuses(main_status_id)
}
}
\usage{
df_main_status
}
\description{
Example package dataset storing the result of rtweet::lookup_statuses(main_status_id)
}
\keyword{datasets}
