% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_retweets}
\alias{df_retweets}
\title{Example package dataset storing the result of rtweet::get_retweets() for multiple status_ids}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 90 columns.
}
\usage{
df_retweets
}
\description{
Example package dataset storing the result of rtweet::get_retweets() for multiple status_ids
}
\examples{
\dontrun{
  df_retweets <- tweet_ids \%>\% map_dfr(~rtweet::get_retweets(.x))
}
}
\keyword{datasets}
