% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_tls}
\alias{df_tls}
\title{Example package dataset storing the result of rtweet::get_timelines(tree_ids)}
\format{
A data frame with 1 row and 90 variables:
\describe{
  df_tls <- rtweet::get_timelines(tree_ids)
}
}
\usage{
df_tls
}
\description{
Example package dataset storing the result of rtweet::get_timelines(tree_ids)
}
\keyword{datasets}
