% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_tree}
\alias{df_tree}
\title{Example package dataset storing the result of search_tree(main_status_id)}
\format{
A data frame with 1 row and 91 variables:
\describe{
  df_tree <- search_tree(main_status_id)
}
}
\usage{
df_tree
}
\description{
Example package dataset storing the result of search_tree(main_status_id)
}
\keyword{datasets}
