% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweettree_tbl_graph.R
\name{find_connections_rec}
\alias{find_connections_rec}
\title{Recursively return all status ids of a tweet and its replies}
\usage{
find_connections_rec(df_tree, df0)
}
\arguments{
\item{df_tree}{data frame returned by `rtweet::search_tweets2()`.}

\item{df0}{data frame returned by `rtweet::lookup_statuses()`.}
}
\value{
Edges data frame that can be plotted by tidygraph & ggraph
}
\description{
Recursively return all status ids of a tweet and its replies
}
\examples{
main_status_id <- "1438481824922181635"
\dontrun{
df_main_status <- rtweet::lookup_statuses(main_status_id)
df_tree <- search_tree(main_status_id)
tree_ids <- df_tree$user_id \%>\% unique()
df_tls <- scrape_timelines(tree_ids)
df0 <- df_main_status \%>\%
                       dplyr::filter(status_id == main_status_id) \%>\%
                       dplyr::select(to = status_id, user_id) \%>\%
                       dplyr::mutate(from = "root", type = "root")
tweet_edges <-
find_connections_rec(dplyr::bind_rows(df_tree, df_tls), df0)
}
}
