% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweettree_data.R
\name{rtweettree_data}
\alias{rtweettree_data}
\alias{rtweettree_data.character}
\alias{rtweettree_data.list}
\alias{rtweettree_data.rtweettree_data}
\title{Scrape rtweettree dataframes from rtweet status_id}
\usage{
rtweettree_data(x, ...)

\method{rtweettree_data}{character}(x, new_tls = TRUE, ...)

\method{rtweettree_data}{list}(x, ...)

\method{rtweettree_data}{rtweettree_data}(x, ...)
}
\arguments{
\item{x}{rtweet status_id}

\item{...}{for the moment not used}
}
\value{
object of class rtweettree_data: a list containing the objects df_main_status, df_tree, df_tls, df_favs & df_retweets (see `vignette("visualize_tree", package = "rtweettree")`)
}
\description{
Scrape rtweettree dataframes from rtweet status_id
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Scrape the rtweettree data from a Twitter status id.

\item \code{list}: Add rtweettree_data class to list.

\item \code{rtweettree_data}: Return rtweettree_data object as is.
}}

\examples{
\dontrun{
l <- rtweettree_data("1438481824922181635")
l
}
}
