% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweettree_data.R
\name{rtweettree_data}
\alias{rtweettree_data}
\title{Scrape rtweettree dataframes from rtweet status_id}
\usage{
rtweettree_data(x, ...)
}
\arguments{
\item{x}{rtweet status_id}

\item{...}{for the moment not used}
}
\value{
object of class rtweettree_data: a list containing the objects df_main_status, df_tree, df_tls, df_favs & df_retweets (see `vignette("visualize_tree", package = "rtweettree")`)
}
\description{
Scrape rtweettree dataframes from rtweet status_id
}
\examples{
\dontrun{
l <- rtweettree_data("1438481824922181635")
l
}
}
