% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_timelines}
\alias{scrape_timelines}
\title{Scrape the timelines of a tree scraped by `search_tree()`}
\usage{
scrape_timelines(tree_ids, main_status_id)
}
\arguments{
\item{tree_ids}{`user_id`s of a tree scraped by `search_tree()`.}

\item{main_status_id}{status id of the root tweet.}
}
\value{
Dataframe of all timelines of all `tree_ids`.
}
\description{
Scrape the timelines of a tree scraped by `search_tree()`
}
\examples{
\dontrun{
main_status_id <- "1234620900386975744"
df_main_status <- rtweet::lookup_statuses(main_status_id)
df_tree <- search_tree(main_status_id)
tree_ids <- df_tree$user_id \%>\% unique()
df_tls <- scrape_timelines(tree_ids, main_status_id)
}
}
